+++ test_dcm_job_processor/test_components/test_service_adapter/test_import_ies.py
@@ -72,7 +72,6 @@ def _report(url, token, request_body):
                     "path": "ie/4a814fe6-b44e-4546-95ec-5aee27cc1d8c",
                     "sourceIdentifier": "test:oai_dc:f50036dd-b4ef",
                     "fetchedPayload": True,
-                    "logId": "0@demo-plugin",
                     "IPIdentifier": None
                 }
             },