+++ test_dcm_job_processor/test_components/test_service_adapter/test_validation_payload.py
@@ -1,12 +1,14 @@
 """
-Test module for the `ServiceAdapter` associated with `Stage.IMPORT_IPS`.
+Test module for the `ServiceAdapter` associated with `Stage.VALIDATION_PAYLOAD`.
 """
 
 import pytest
 from dcm_common.services import APIResult
 
 from dcm_job_processor.models import Stage
-from dcm_job_processor.components.service_adapter import ImportIPsAdapter
+from dcm_job_processor.components.service_adapter import (
+    ValidationPayloadAdapter
+)
 
 
 @pytest.fixture(name="port")
@@ -21,20 +23,22 @@ def _url(port):
 
 @pytest.fixture(name="adapter")
 def _adapter(url):
-    return ImportIPsAdapter(url)
+    return ValidationPayloadAdapter(url)
 
 
 @pytest.fixture(name="target")
 def _target():
     return {
-        "path": "ips"
+        "path": "ip/59438ebf-75e0-4345-8d6b-132a57e1e4f5"
     }
 
 
 @pytest.fixture(name="request_body")
 def _request_body():
     return {
-        "import": {}
+        "validation": {
+            "plugins": {"request-0": {"plugin": "plugin-0", "args": {}}}
+        }
     }
 
 
@@ -62,45 +66,41 @@ def _report(url, token, request_body):
             "EVENT": [
                 {
                     "datetime": "2024-08-09T12:15:10+00:00",
-                    "origin": "Import Module",
+                    "origin": "Object Validator",
                     "body": "Some event"
                 },
             ]
         },
         "data": {
             "success": True,
-            "IPs": {
-                "ips/a": {
-                    "path": "ips/a",
-                    "logId": ["ips/a@object_validator"],
-                }
-            },
+            "valid": True,
+            "details": {}
         }
     }
 
 
 @pytest.fixture(name="report_fail")
 def _report_fail(report):
-    report["data"]["success"] = False
+    report["data"]["valid"] = False
     return report
 
 
-@pytest.fixture(name="import_module")
-def _import_module(port, token, report, run_service):
+@pytest.fixture(name="object_validator")
+def _object_validator(port, token, report, run_service):
     run_service(
         routes=[
-            ("/import/internal", lambda: (token, 201), ["POST"]),
+            ("/validate", lambda: (token, 201), ["POST"]),
             ("/report", lambda: (report, 200), ["GET"]),
         ],
         port=port
     )
 
 
-@pytest.fixture(name="import_module_fail")
-def _import_module_fail(port, token, report_fail, run_service):
+@pytest.fixture(name="object_validator_fail")
+def _object_validator_fail(port, token, report_fail, run_service):
     run_service(
         routes=[
-            ("/import/internal", lambda: (token, 201), ["POST"]),
+            ("/validate", lambda: (token, 201), ["POST"]),
             ("/report", lambda: (report_fail, 200), ["GET"]),
         ],
         port=port
@@ -109,13 +109,13 @@ def _import_module_fail(port, token, report_fail, run_service):
 
 def fix_report_args(info: APIResult, target) -> None:
     """Fixes args in report (missing due to faked service)"""
-    info.report["args"]["import"]["target"] = target
+    info.report["args"]["validation"]["target"] = target
 
 
 def test_run(
-    adapter: ImportIPsAdapter, request_body, target, report, import_module
+    adapter: ValidationPayloadAdapter, request_body, target, report, object_validator
 ):
-    """Test method `run` of `ImportIPsAdapter`."""
+    """Test method `run` of `ValidationPayloadAdapter`."""
     adapter.run(request_body, target, info := APIResult())
     fix_report_args(info, target)
     assert info.completed
@@ -124,9 +124,9 @@ def test_run(
 
 
 def test_run_fail(
-    adapter: ImportIPsAdapter, request_body, target, report_fail, import_module_fail
+    adapter: ValidationPayloadAdapter, request_body, target, report_fail, object_validator_fail
 ):
-    """Test method `run` of `ImportIPsAdapter`."""
+    """Test method `run` of `ValidationPayloadAdapter`."""
     adapter.run(request_body, target, info := APIResult())
     fix_report_args(info, target)
     assert info.completed
@@ -135,80 +135,70 @@ def test_run_fail(
 
 
 def test_success(
-    adapter: ImportIPsAdapter, request_body, target, import_module
+    adapter: ValidationPayloadAdapter, request_body, target, object_validator
 ):
-    """Test property `success` of `ImportIPsAdapter`."""
+    """Test property `success` of `ValidationPayloadAdapter`."""
     adapter.run(request_body, target, info := APIResult())
     assert adapter.success(info)
 
 
 def test_success_fail(
-    adapter: ImportIPsAdapter, request_body, target, import_module_fail
+    adapter: ValidationPayloadAdapter, request_body, target, object_validator_fail
 ):
-    """Test property `success` of `ImportIPsAdapter`."""
+    """Test property `success` of `ValidationPayloadAdapter`."""
     adapter.run(request_body, target, info := APIResult())
     assert not adapter.success(info)
 
 
 def test_export_records(
-    adapter: ImportIPsAdapter, request_body, target, import_module
+    adapter: ValidationPayloadAdapter, request_body, target, object_validator
 ):
-    """Test method `export_records` of `ImportIPsAdapter`."""
+    """Test method `export_records` of `ValidationPayloadAdapter`."""
     adapter.run(request_body, target, info := APIResult())
+    fix_report_args(info, target)
     records = adapter.export_records(info)
     assert len(records) == 1
     ip_id = list(records)[0]
-    assert Stage.IMPORT_IPS in records[ip_id].stages
-    assert records[ip_id].stages[Stage.IMPORT_IPS].report == info.report
+    assert Stage.VALIDATION_PAYLOAD in records[ip_id].stages
+    assert (
+        records[ip_id].stages[Stage.VALIDATION_PAYLOAD].report == info.report
+    )
 
 
 def test_export_records_fail(
-    adapter: ImportIPsAdapter, request_body, target, import_module_fail
+    adapter: ValidationPayloadAdapter, request_body, target, object_validator_fail
 ):
-    """Test method `export_records` of `ImportIPsAdapter`."""
+    """Test method `export_records` of `ValidationPayloadAdapter`."""
     adapter.run(request_body, target, info := APIResult())
+    fix_report_args(info, target)
     records = adapter.export_records(info)
     assert len(records) == 1
     ip_id = list(records)[0]
-    assert Stage.IMPORT_IPS in records[ip_id].stages
-    assert records[ip_id].stages[Stage.IMPORT_IPS].report == info.report
+    assert Stage.VALIDATION_PAYLOAD in records[ip_id].stages
+    assert (
+        records[ip_id].stages[Stage.VALIDATION_PAYLOAD].report == info.report
+    )
 
 
 def test_export_target(
-    adapter: ImportIPsAdapter, request_body, target, import_module
+    adapter: ValidationPayloadAdapter, request_body, target, object_validator
 ):
     """
-    Test method `export_target` of `ImportIPsAdapter`.
+    Test method `export_target` of `ValidationPayloadAdapter`.
     """
     adapter.run(request_body, target, info := APIResult())
+    fix_report_args(info, target)
     target = adapter.export_target(info)
-    assert target["path"] == info.report["data"]["IPs"]["ips/a"]["path"]
+    assert target == info.report["args"]["validation"]["target"]
 
 
-def test_export_target_multiple(
-    adapter: ImportIPsAdapter, request_body, target, import_module
+def test_export_target_fail(
+    adapter: ValidationPayloadAdapter, request_body, target, object_validator_fail
 ):
     """
-    Test method `export_target` of `ImportIPsAdapter` for two records.
+    Test method `export_target` of `ValidationPayloadAdapter`.
     """
     adapter.run(request_body, target, info := APIResult())
-    info.report["data"]["IPs"]["ips/b"] = (
-        info.report["data"]["IPs"]["ips/a"].copy()
-    )
-    info.report["data"]["IPs"]["ips/b"]["path"] = "ips/b"
-    assert (
-        adapter.export_target(info)["path"] in [
-            ip["path"] for ip in info.report["data"]["IPs"].values()
-        ]
-    )
-
-
-def test_export_target_none(
-    adapter: ImportIPsAdapter, request_body, target, import_module
-):
-    """
-    Test method `export_target` of `ImportIPsAdapter` for no records.
-    """
-    adapter.run(request_body, target, info := APIResult())
-    info.report["data"]["IPs"] = {}
-    assert adapter.export_target(info) is None
+    fix_report_args(info, target)
+    target = adapter.export_target(info)
+    assert target is None