+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## [1.0.0] - 2025-02-14
+
+### Changed
+
+- **Breaking:** migrate to latest Import Module API (v6)
+- **Breaking:** migrate to latest Object Validator API (v5)
+- **Breaking:** migrate to latest IP Builder API (v4)
+
 ## [0.1.2] - 2024-11-21
 
 ### Changed